backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
APScheduler==3.10.4
python-telegram-bot==20.7
python-dotenv==1.0.0
gunicorn==21.2.0
